fixup! Bug 1517: Reduce precision of time for Javascript.

+++ dom/events/KeyboardEvent.h
@@ -42,7 +42,7 @@ public:
     return KeyboardEventBinding::Wrap(aCx, this, aGivenProto);
   }
 
-  uint64_t TimeStamp() const
+  double TimeStamp() const
   {
     return (mEvent->time / 250)*250;
   }
